% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Calculate the Maximum Likelihood Estimate for Bernoulli Distribution}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A vector of binary outcomes (0s and 1s).}
}
\value{
The maximum likelihood estimate (MLE) for the parameter of the Bernoulli distribution based on grid search.
}
\description{
This function calculates the MLE for Bernoulli Distribution using Grid search.
}
\examples{
data <- c(1, 1, 0, 1, 1, 1, 1)
logLikBernoulli(data)
}
