% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Calculate Minimum Sample Size for Two-sample t-test}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{Sample data of first group.}

\item{x2}{Sample data of second group.}
}
\value{
The minimum sample size required for the t-test. for a two sample t-test, the returned value is sample size for each group.
}
\description{
This function calculates the minimum sample size required for a two-sided one or two-sample t-test.
}
\examples{
x1 = c(162, 123.5, 133.6, 178.2, 100, 125.6, 110, 105, 106)
x2 = c(99, 120.8, 105.3, 88, 87, 125, 92, 105, 107.8)
minimumN(x1, x2)
x1 = c(1, 2, 3, 4, 5, 6, 7, 8, 9)
minimumN(x1)
}
