% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{Approximate data using principal components}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{A numeric vector or dataframe containing data.}

\item{npc}{An integer specifying the number of principal components.}
}
\value{
A numeric vector or dataframe containing the approximation to the data.
}
\description{
This function returns an approximation to the data based on the specified number of principal components.
}
\examples{
set.seed(123)
x = tibble::tibble(x = c(1, 2, 3, 0, 7, 8, 9), y=c(21, 85.0, 13.2, 21, 99, 61, 10.8))
npc = 2
approximation <- pcApprox(x, npc)
}
