% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize column names}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A dataframe.}
}
\value{
A dataframe with standardized column names.
}
\description{
This function standardizes the column names of a dataframe by converting them to lower camel case.
}
\examples{
data <- data.frame("First Name" = c("Ram", "Smith"), "birth_month" = c("June", "March"))
standardized_data <- standardizeNames(data)
colnames(standardized_data)
# Output should be c("firstName", "birthMonth")
}
