% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Plots a Survival Curve}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{A vector indicating the status of each individual (1 for event occurrence, 0 for censored).}

\item{time}{A vector of survival times.}
}
\description{
This function plots the survival curve using the provided survival times and event status.
}
\details{
Estimates survival using Kaplan-Meier.

This function requires the ggplot2 and survival packages to be installed.
}
\examples{
# Example usage
status <- c(1, 1, 0, 1, 0)
time <- c(2, 4, 6, 8, 10)
survCurv(status, time)

}
