% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Transforms a scaled Vector back to it's original form.}
\usage{
unscale(x)
}
\arguments{
\item{x}{A numeric vector that has been scaled.}
}
\value{
The unscaled vector.
}
\description{
This function unscales a vector that has been previously scaled using the `scale()` function.
}
\details{
This function retrieves the mean and standard deviation used for scaling from the attributes
of the input vector and then unscales the vector accordingly.
}
\examples{
# Example usage
# Example 1
scaled_vector = scale(1:10)
unscaled_vector = unscale(scaled_vector)
unscaled_vector

# Example 2
x = c(2, 3, 4, 5)
attr(x, "scaled:center") = mean(x)
attr(x, "scaled:scale") = sd(x)
unscaled_vector = unscale(x)
unscaled_vector
}
